+++ project_composer/marker.py
@@ -1,13 +1,13 @@
 
 class EnabledApplicationMarker:
     """
-    Empty class to mark a settings class mixin as enabled to load for composition.
+    Empty class to mark another class as enabled to load for composition.
 
     This class won't never ever introduce anything else than marker attribute
     ``_ENABLED_COMPOSABLE_APPLICATION`` since the class may be herited by any kind of
     class we don't want to pollute.
 
     Only purpose of attribute ``_ENABLED_COMPOSABLE_APPLICATION`` is to mark it for
-    inspection from composer.
+    inspection from composer, don't change it.
     """
     _ENABLED_COMPOSABLE_APPLICATION = True